Equity/Portfolio, Portfolio Import, other basic things
Added the basic functions/types for working with equities and
portfolio(s). Also added a function to import a portfolio from a CSV
file downloaded from Fidelity's website.

Some auxillary modules were implemented as well.
- Static Mem Cache: A module for managing an array of a single type.
  This will be used to allocate and free equities and events. I'm trying
  to avoid dynamic memory allocation in this project.
- Parts of the event loop module: I'm not exactly sure what this will
  look like yet, but added an event queue. After initialization, the
  software will be event driven.

This is just a skeleton and things (especially the things for the event
loop) are bound to change. The next step will be adding the code to
fetch current pricing information of equities. Once this is done, the
event loop can be fleshed out.