Declare needed CURL callbacks and timestamp_ms function
There are three major CURL callback functions we must define to use the
mutli interface.
1. Timeout: CURL telling our application it needs to run something in X
   ms.
2. Socket: CURL telling our application to listen for activity on socket
   X.
3. Write data: CURL telling our application to copy data into our
   buffers.
I implemented the timeout one because it is very simple, but I'm not
sure what the other two will need to look like (in terms of the user
pointer) so I'm holding off. Before I start implementing these two, I
need to figure out the following:
1. How do we keep track of which fds epoll should poll? What will the
   data structure look like.
2. How do we want to allocate buffers for the CURL write callback?
   Should we use static buffers or turn to the dark side and just use
   malloc?
3. How should we keep track of in flight event io? What will the data
   structure look like.
Once these are answered, it should be fairly straightforward to
implement the callbacks and event loop.

I also added a function for getting a timestamp in milliseconds. Ordinarily,
this would not be noteworthy but I had to do something bad to accomplish
it. I'll probably describe the problem better in the README when I get
time but I'll briefly describe it now.

I like using -std=c11 and -Wpedantic because it makes the code more
portable. But, as I discovered this gloomy morning, sometimes you need to
get your hands diry. The standard C time.h functions do not include a
function that returns an NTP (and other time changes) agnostic time.
Generally this is ok, but if you are doing any sort of scheduling, it
can lead to bugs if the time "randomly" jitters.

POSIX defines a function, clock_gettime, that can return a time not
affected by these changes. It does not correspond to epoch time, but
that generally doesn't matter for scheduling based on relative time
(schedule this in X ms). By passing CLOCK_BOOTTIME into clock_gettime
we can retrieve a time that is not affected by these jitters caused by
NTP and includes suspend time. CLOCK_MONOTONIC is similar but does not
include suspend time.

It is noteworthy that CLOCK_BOOTTIME is still affected by NTP, but the
change is smoothed out over time.